+++ packages/lib/test/actionMiddlewares/onActionMiddleware.test.ts
@@ -12,7 +12,6 @@ import {
   onActionMiddleware,
   prop,
 } from "../../src"
-import "../commonSetup"
 import { autoDispose } from "../utils"
 
 @model("P2")