+++ packages/lib/test/redux/redux.test.ts
@@ -10,7 +10,6 @@ import {
   ReduxMiddleware,
   SnapshotOutOfModel,
 } from "../../src"
-import "../commonSetup"
 import { autoDispose } from "../utils"
 
 @model("P")