+++ packages/lib/test/ref/rootRef.test.ts
@@ -24,7 +24,6 @@ import {
   runUnprotected,
   undoMiddleware,
 } from "../../src"
-import "../commonSetup"
 import { autoDispose } from "../utils"
 
 @model("Country")